// src/types/vue-logics.d.ts
  interface Position {
    top: number;
    left: number;
    scrollTop: number;
    scrollLeft: number;
  }
// ...
export namespace BlockElement {
  interface Position {
    top: number;
    left: number;
    height: number;
    width: number;
  }
  export interface BlockElementOptions {
    window?: Window;
    parent?: number;
    childWidth?: number;
  }
  export interface BlockElementConstructor {
    new(id: number, node: HTMLDivElement, options: BlockElementOptions);
    static find: (id: number, options: BlockElementOptions) => BlockElementInterface | null;
    static fromElement: (node: HTMLDivElement, options: BlockElementOptions) => BlockElementInterface | null;
  }
  export interface BlockElementInterface {
    id: number;
    node: HTMLDivElement;
    window: Window;

    position: () => Position;
    styles: (styles: any) => CSSStyleDeclaration;
    arrow: () => any;
  }
}

export namespace ArrowElement {
  export interface Position {
    top: number;
    left: number;
  }
  export interface ArrowElementConstructor {
    new(blockElement: BlockElement.BlockElementInterface, node: HTMLDivElement);
    static find: (blockElement: BlockElement.BlockElementInterface) => ArrowElementInterface | null;
  }
  export interface ArrowElementInterface {
    blockElement: BlockElement.BlockElementInterface;
    node: HTMLDivElement;
    window: Window;
    document: HTMLDivElement;

    html: (html?: string) => string;
    position: () => Position;
    styles: (styles: any) => CSSStyleDeclaration;
    remove: () => void;
  }
}